Move keystore files into their own module
