+++ account_manager/Cargo.toml
@@ -32,5 +32,6 @@ rust-crypto = "0.2.36"
 hex = "0.4.0"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
+uuid = { version = "0.8", features = ["serde", "v4"] }
