+++ account_manager/src/main.rs
@@ -7,6 +7,7 @@ use types::test_utils::generate_deterministic_keypair;
 use types::DepositData;
 use validator_client::Config as ValidatorClientConfig;
 mod deposit;
+mod keystore;
 pub const DEFAULT_DATA_DIR: &str = ".lighthouse-validator";
 pub const CLIENT_CONFIG_FILENAME: &str = "account-manager.toml";
